# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Velora Wear API
  description: |
    RESTful API for Velora Wear - A sustainable fashion e-commerce platform.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## User Roles
    - **customer**: Regular users who can browse and purchase products
    - **admin**: Administrators with access to management endpoints
    
  version: 1.0.0
  contact:
    name: Velora Wear API Support
    email: support@velorawear.com
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.velorawear.com/api
    description: Production server
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve all products with optional filters and sorting
      operationId: getProducts
      parameters:
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
            enum: [Casual Wear, Streetwear, Essentials, Limited Edition]
        - name: gender
          in: query
          description: Filter by gender
          schema:
            type: string
            enum: [Men, Women, Unisex]
        - name: search
          in: query
          description: Search in product name and description
          schema:
            type: string
        - name: sort
          in: query
          description: Sort products
          schema:
            type: string
            enum: [price-low, price-high, newest, latest]
        - name: featured
          in: query
          description: Filter featured products
          schema:
            type: boolean
        - name: newArrivals
          in: query
          description: Filter new arrivals
          schema:
            type: boolean
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags:
        - Products
      summary: Create a product
      description: Create a new product (Admin only)
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve a single product by its ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      tags:
        - Products
      summary: Update a product
      description: Update an existing product (Admin only)
      operationId: updateProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags:
        - Products
      summary: Delete a product
      description: Delete an existing product (Admin only)
      operationId: deleteProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Product removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  # Order Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    Product:
      type: object
      properties:
        _id:
          type: string
          example: 60d5ec49f1b2c72b8c8e4a1c
        name:
          type: string
          example: Classic Cotton Tee
        description:
          type: string
          example: Sustainable organic cotton t-shirt
        price:
          type: number
          example: 45.00
        category:
          type: string
          enum: [Casual Wear, Streetwear, Essentials, Limited Edition]
          example: Essentials
        gender:
          type: string
          enum: [Men, Women, Unisex]
          example: Unisex
        sizes:
          type: array
          items:
            type: string
            enum: [XS, S, M, L, XL, XXL]
          example: [S, M, L, XL]
        colors:
          type: array
          items:
            type: string
          example: [Black, White, Gray]
        images:
          type: array
          items:
            type: string
          example: ["https://example.com/image1.jpg"]
        stock:
          type: integer
          example: 100
        isFeatured:
          type: boolean
          example: true
        isNewArrival:
          type: boolean
          example: false
        rating:
          type: number
          minimum: 0
          maximum: 5
          example: 4.5
        numReviews:
          type: integer
          example: 12
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required:
        - name
        - description
        - price
        - category
        - gender
        - stock
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0
        category:
          type: string
          enum: [Casual Wear, Streetwear, Essentials, Limited Edition]
        gender:
          type: string
          enum: [Men, Women, Unisex]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
          minimum: 0
        isFeatured:
          type: boolean
        isNewArrival:
          type: boolean
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Error message
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              location:
                type: string
  responses:
...
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Not authorized, no token
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Not authorized as admin
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Resource not found
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Server error